refactor: refactor form components into separate files and improve styling
