+++ components/index.tsx
@@ -1,6 +1,9 @@
 export * from './action';
 export * from './button';
-export * from './form';
+export * from './input';
+export * from './checkbox';
+export * from './pill';
+export * from './textarea';
 export * from './paper';
 export * from './spinner';
 export * from './theme';
