+++ components/input/Input.stories.tsx
@@ -1,35 +1,39 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, text, boolean } from '@storybook/addon-knobs';
-import {
-  Input, Checkbox, Pill
-} from './index';
 
-const stories = storiesOf('Form', module);
+import { Input } from './index';
+
+const stories = storiesOf('Input', module);
+
 stories.addDecorator(withKnobs);
 stories.addDecorator(withInfo);
 stories.addParameters({ info: { inline: true } });
 
 stories.add('Input', () => {
   const placeholder: string = text('Placeholder', 'Email');
+
   return <Input placeholder={placeholder} onChange={action('onChange')} />;
 });
 
 stories.add('Input with label', () => {
   const placeholder: string = text('Label', 'Email');
+
   return <Input label={placeholder} onChange={action('onChange')} />;
 });
 
 stories.add('Input Error', () => {
   const placeholder: string = text('Placeholder', 'Email');
   const errored: boolean = boolean('Error', true);
+
   return <Input hasError={errored} placeholder={placeholder} onChange={action('onChange')} />;
 });
 
 stories.add('Input with Icon', () => {
   const placeholder: string = text('Placeholder', 'Email');
+
   return (
     <div>
       <Input placeholder={placeholder} onChange={action('onChange')} icon={<img width="100%" src="/copy.svg" alt="copy" />} />
@@ -44,21 +48,3 @@ stories.add('Input disabled', () => (
     <Input value="readonly" onChange={action('onChange')} readOnly />
   </div>
 ));
-
-stories.add('Pill', () => (
-  <div>
-    <Pill>Testpillcomponent One</Pill>
-    <Pill>Testpillcomponent Two</Pill>
-  </div>
-));
-
-const CheckBoxStory = () => {
-  const [checked, check] = useState(true);
-  return (
-    <Checkbox checked={checked} onChange={() => { check(!checked); action('Checked')(checked); }} id="terms_of_service">
-      <span>I agree to the Terms of Service</span>
-    </Checkbox>
-  );
-};
-
-stories.add('Checkbox', () => <CheckBoxStory />);
